Various minor bug fixes for 3.0rc14 release.
	modified:   idas.chox.core/src/main/java/idas/chox/core/model/HireMonitoringDetail.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/bre/rules/LabourCostBusinessRule.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/bre/util/ClaimCalcHelper.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/bre/util/mathHelper.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/viewdata/ClaimFileReportData.java
	modified:   idas.chox.web/src/main/resources/reports/template_InvoiceReport.xls

+++ idas.chox.service/src/main/java/idas/chox/service/reports/viewdata/ClaimFileReportData.java
@@ -128,7 +128,7 @@ public class ClaimFileReportData {
     private String hireMonRepairCompletionDate;
     private String hireMonIME;
     private BigDecimal hireMonLabourRate;
-    private Integer hireMonLabourHours;
+    private BigDecimal hireMonLabourHours;
     private BigDecimal hireMonTotalLabourCost;
     private String hireMonNonProvisionReason;
     private String hireVehicleManufacturer;
@@ -1034,11 +1034,11 @@ public class ClaimFileReportData {
         this.hireMonInspectionDate = hireMonInspectionDate;
     }
 
-    public Integer getHireMonLabourHours() {
+    public BigDecimal getHireMonLabourHours() {
         return hireMonLabourHours;
     }
 
-    public void setHireMonLabourHours(Integer hireMonLabourHours) {
+    public void setHireMonLabourHours(BigDecimal hireMonLabourHours) {
         this.hireMonLabourHours = hireMonLabourHours;
     }
